% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betting_tools.R
\name{calculateKellyStake}
\alias{calculateKellyStake}
\title{Calculates a single Kelly stake}
\usage{
calculateKellyStake(
  expected,
  payout,
  kelly_multiplier = 1,
  expected_odds = "prob",
  payout_odds = "dec"
)
}
\arguments{
\item{expected}{A numerical value representing the odds of the expected payout.}

\item{payout}{A numerical value representing the odds of the actual payout.}

\item{kelly_multiplier}{A numerical value representing the kelly multiplier of the bet.
The default value of the Kelly multiplier is 1.}

\item{expected_odds}{A string representing the odds format of the expected payout.
It is either "prob", "dec", or "us". The default odds format is in implied probability.}

\item{payout_odds}{A string representing the odds format of the actual payout.
It is either "prob", "dec", or "us". The default odds format is in decimal odds.}
}
\value{
The single Kelly stake.
}
\description{
Returns the single Kelly stake. This is the percentage of one's bankroll
one should bet to maximize the expected growth of one's bankroll on a single bet.
This is known as the single Kelly stake.
}
\examples{
calculateKellyStake(0.41, 2.56)

calculateKellyStake(-120, 150, expected_odds = "us", payout_odds = "us")

calculateKellyStake(0.70, -150, kelly_multiplier = 0.1, payout_odds = "us")

calculateKellyStake(0.26, -110, payout_odds = "us")
}
