% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betting_tools.R
\name{calculateZeroVigProb}
\alias{calculateZeroVigProb}
\title{Calculate zero-vig implied probabilities}
\usage{
calculateZeroVigProb(lines, precision = 4)
}
\arguments{
\item{lines}{A vector representing a vector of lines in American odds format.}

\item{precision}{A numerical value representing the precision. The default precision
is set to 4 digits.}
}
\value{
A vector representing the zero-vig implied probabilities.
}
\description{
Returns a vector of the zero-vig implied probabilities
}
\examples{
calculateZeroVigProb(c(200, -180, -450, 800))

calculateZeroVigProb(-237)

calculateZeroVigProb(-237, precision = 7)
}
