% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds_converter.R
\name{Dec2All}
\alias{Dec2All}
\title{Convert Decimal to American and Implied odds}
\usage{
Dec2All(decimal, precision = 4)
}
\arguments{
\item{precision}{A numerical value representing the precision. The default precision
is set to 4 digits.}

\item{american}{A vector of Decimal odds.}
}
\value{
A tibble of American, Decimal, and Implied odds
}
\description{
Returns a tibble where each row has American, Decimal, and Implied
odds
}
\examples{
Dec2All(c(3.17, 2.14, 2.01, 1.67))
}
