% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds_converter.R
\name{Dec2Implied}
\alias{Dec2Implied}
\title{Convert Decimal to Implied odds}
\usage{
Dec2Implied(decimal, precision = 4)
}
\arguments{
\item{decimal}{A vector of Decimal odds.}

\item{precision}{A numerical value representing the precision. The default precision
is set to 4 digits.}
}
\value{
A vector of Implied odds.
}
\description{
Returns a vector of Implied odds
}
\examples{
Dec2Implied(c(3.17, 2.14, 2.01, 1.67))
}
