% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds_converter.R
\name{Dec2US}
\alias{Dec2US}
\title{Convert Decimal to American odds}
\usage{
Dec2US(decimal, precision = 4)
}
\arguments{
\item{decimal}{A vector of Decimal odds.}

\item{precision}{A numerical value representing the precision. The default precision
is set to 4 digits.}
}
\value{
A vector of American odds.
}
\description{
Returns a vector of American odds
}
\examples{
Dec2US(c(3.17, 2.14, 2.01, 1.67))
}
