% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds_converter.R
\name{Implied2All}
\alias{Implied2All}
\title{Convert Implied to Decimal and Implied odds}
\usage{
Implied2All(implied, precision = 4)
}
\arguments{
\item{implied}{A vector of Implied odds.}

\item{precision}{A numerical value representing the precision. The default precision
is set to 4 digits.}
}
\value{
A tibble of American, Decimal, and Implied odds
}
\description{
Returns a tibble where each row has American, Decimal, and Implied
odds
}
\examples{
Implied2All(c(.34, .54, .88, .12))
}
