% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds_converter.R
\name{Implied2Dec}
\alias{Implied2Dec}
\title{Convert Implied to Decimal odds}
\usage{
Implied2Dec(implied, precision = 2)
}
\arguments{
\item{precision}{A numerical value representing the precision. The default precision
is set to 4 digits.}

\item{decimal}{A vector of Implied odds.}
}
\value{
A vector of Decimal odds.
}
\description{
Returns a vector of Decimal odds
}
\examples{
Implied2Dec(c(.34, .54, .88, .12))
}
