% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds_converter.R
\name{Implied2US}
\alias{Implied2US}
\title{Convert Implied to American odds}
\usage{
Implied2US(implied, precision = 4)
}
\arguments{
\item{precision}{A numerical value representing the precision. The default precision
is set to 4 digits.}

\item{decimal}{A vector of Implied odds.}
}
\value{
A vector of American odds.
}
\description{
Returns a vector of American odds
}
\examples{
Implied2US(c(.34, .54, .88, .12))
}
