% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_data.R
\docType{data}
\name{nba_odds_2020}
\alias{nba_odds_2020}
\title{2019-2020 NBA Historical Vegas Lines}
\format{\describe{
   \item{date}{Date of NBA game}
   \item{game}{Game number}
   \item{away_name}{Away team full name}
   \item{home_name}{Home team full name}
   \item{away_score}{Away team score}
   \item{home_score}{Home team score}
   \item{away_open_ml}{Away team opening moneyline in American odds format}
   \item{home_open_ml}{Home team opening moneyline in American odds format}
   \item{away_close_ml}{Away team closing moneyline in American odds format}
   \item{home_close_ml}{Home team closing moneyline in American odds format}
}}
\source{
\url{https://www.sportsbookreviewsonline.com/scoresoddsarchives/nba/nbaoddsarchives.htm}
}
\usage{
nba_odds_2020
}
\description{
Dataset containing historical 2019-2020 NBA Vegas lines on moneylines, point totals,
and point spreads.
}
\keyword{datasets}
