% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds_converter.R
\name{US2All}
\alias{US2All}
\title{Convert American to Decimal and Implied odds}
\usage{
US2All(american, precision = 4)
}
\arguments{
\item{american}{A vector of American odds.}

\item{precision}{A numerical value representing the precision. The default precision
is set to 4 digits.}
}
\value{
A tibble of American, Decimal, and Implied odds
}
\description{
Returns a tibble where each row has American, Decimal, and Implied
odds
}
\examples{
US2All(c(-250, 600, 137, -110))
}
