% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds_converter.R
\name{US2Dec}
\alias{US2Dec}
\title{Convert American to Decimal odds}
\usage{
US2Dec(american, precision = 2)
}
\arguments{
\item{american}{A vector of American odds.}

\item{precision}{A numerical value representing the precision. The default precision
is set to 4 digits.}
}
\value{
A vector of Decimal odds.
}
\description{
Returns a vector of Decimal odds
}
\examples{
US2Dec(c(-250, 600, 137, -110))
}
