% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds_converter.R
\name{US2Fair}
\alias{US2Fair}
\title{Calculate fair lines from a pair of American odds}
\usage{
US2Fair(pair, precision = 4)
}
\arguments{
\item{pair}{A vector representing the two outcomes in American odds format.}

\item{precision}{A numerical value representing the precision. The default precision
is set to 4 digits.}
}
\value{
A vector representing the fair line
}
\description{
Returns a vector of fair lines
}
\examples{
US2Fair(c(265, -375))

US2Fair(c(150, -200))
}
