% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds_converter.R
\name{US2Implied}
\alias{US2Implied}
\title{Convert American to Implied odds}
\usage{
US2Implied(american, precision = 4)
}
\arguments{
\item{american}{A vector of American odds.}

\item{precision}{A numerical value representing the precision. The default precision
is set to 4 digits.}
}
\value{
A vector of Implied odds.
}
\description{
Returns a vector of Implied odds
}
\examples{
US2Implied(c(-250, 600, 137, -110))
}
